+++ tests/mainTest.py
@@ -7,7 +7,9 @@ printLock = threading.Lock()
 
 class Structure:
     def __init__(self):
+        print(f'{self} attempting to initialize lock')
         self.lock = Lock()
+        print(f'{self} attempting to initialize condition')
         self.condition = Condition(self.lock)
         #self.lock.setName('test')
     
@@ -43,17 +45,17 @@ class MyThread(Thread):
             sleep(5)
 
 if __name__ == "__main__":
+    print('starting main')
     structure1 = Structure()
     structure2 = Structure()
     structure3 = Structure()
-
+    print('structures initialized')
     threads = []
-
+    print('initializing threads')
     for i in range(4):
         t = MyThread(structure3,structure2,structure1)
         threads.append(t)
-
+    print('starting threads')
     for t in threads:
         t.start()
-
-    startGraph()
+    print('starting threadmonitor')
