+++ threadmonitor/view/__init__.py
@@ -7,4 +7,6 @@ from . import tk
 activeViews = []
 
 def view_init():
+    
     activeViews.append( tk.setup() )
+    print('view initialized')
